Move cancellable to own utility module
